requirements.txt
base58
dnslib
dnspython
